workshops/o11y-vector-ai/embedding-pipeline/requirements.txt
clickhouse-connect==0.7.0
openai==1.12.0
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
schedule==1.2.0
